ejection: moving files to simplify pruning
