+++ src/lib/components/dropdown/DropdownItem.js
@@ -2,8 +2,7 @@ import React, { useContext } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import ThemeWrapper from '../../utils/ThemeWrapper'
-
-import { DropdownContext } from './DropdownContext'
+import DropdownContext from '../../utils/DropdownContext'
 
 export const DropdownItem = ({
   active,
