+++ src/lib/utils/DropdownContext.js
@@ -1,8 +1,10 @@
 import React from 'react'
 
-export const DropdownContext = React.createContext({
+const DropdownContext = React.createContext({
   toggle: () => {},
   open: false,
   direction: 'down',
   nav: false
 })
+
+export default DropdownContext
