+++ src/lib/utils/Fade.js
